@@ -6,60 +6,54 @@
   "author": "Preston Lee",
   "scripts": {
     "start": "grunt",
-    "postinstall": "typings install"
+    "postinstall": "typings install && node ./ng-add-pug-loader.js"
   },
   "dependencies": {
-    "@angular/animations": "^4.4.6",
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/forms": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/router": "^4.4.6",
-    "@angular/upgrade": "^4.4.6",
-    "@types/core-js": "^0.9.43",
-    "angular-in-memory-web-api": "^0.3.2",
-    "angular2-router-loader": "^0.3.5",
-    "angular2-toaster": "^4.0.1",
+    "@angular/animations": "^7.2.2",
+    "@angular/common": "^7.2.2",
+    "@angular/compiler": "^7.2.2",
+    "@angular/core": "^7.2.2",
+    "@angular/forms": "^7.2.2",
+    "@angular/http": "^7.2.2",
+    "@angular/platform-browser": "^7.2.2",
+    "@angular/platform-browser-dynamic": "^7.2.2",
+    "@angular/router": "^7.2.2",
+    "@angular/upgrade": "^7.2.2",
+    "angular-in-memory-web-api": "^0.8.0",
+    "angular2-toaster": "^7.0.0",
     "angular2-uuid": "^1.1.1",
-    "bootstrap": "^3.3.7",
+    "bootstrap": "^4.2.1",
     "core-js": "^2.5.3",
-    "font-awesome": "^4.7.0",
-    "jquery": "^3.2.1",
-    "ng2-bootstrap": "^1.6.3",
-    "ng2-cable": "^1.1.9",
-    "reflect-metadata": "^0.1.10",
-    "rxjs": "^5.5.6",
-    "webpack": "^2.7.0",
-    "webpack-dev-server": "^2.9.7",
-    "zone.js": "^0.8.18"
+    "jquery": "^3.3.1",
+    "ng-cli-pug-loader": "^0.1.7",
+    "popper.js": "^1.14.6",
+    "rxjs": "^6.3.3",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "concurrently": "^3.5.1",
-    "node-sass": "^4.7.2",
-    "grunt": "^1.0.1",
-    "grunt-cli": "^1.2.0",
-    "grunt-concurrent": "^2.3.1",
-    "grunt-contrib-clean": "^1.1.0",
-    "grunt-contrib-compress": "^1.4.3",
-    "grunt-contrib-concat": "^1.0.1",
-    "grunt-contrib-connect": "^1.0.2",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-cssmin": "^2.2.1",
-    "grunt-contrib-pug": "^1.0.0",
-    "grunt-contrib-uglify": "^3.2.1",
-    "grunt-contrib-watch": "^1.0.0",
-    "grunt-sass": "^2.0.0",
-    "grunt-webpack": "^3.0.2",
-    "install": "^0.10.2",
-    "lite-server": "^2.3.0",
-    "load-grunt-tasks": "^3.5.2",
-    "npm": "^5.6.0",
-    "source-map-loader": "^0.2.3",
-    "ts-loader": "^2.3.7",
-    "typescript": "^2.6.2",
-    "typings": "^2.1.1"
+    "@angular-devkit/build-angular": "^0.12.3",
+    "@angular/cli": "^7.0.0",
+    "@angular/compiler-cli": "^7.0.0",
+    "@angular/platform-server": "^7.0.0",
+    "@types/jasmine": "~3.3.7",
+    "@types/jasminewd2": "^2.0.4",
+    "@types/node": "~10.12.18",
+    "apply-loader": "^2.0.0",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-phantomjs-launcher": "^1.0.2",
+    "lodash": "^4.16.2",
+    "phantomjs-prebuilt": "^2.1.7",
+    "protractor": "~5.4.0",
+    "pug": "^2.0.3",
+    "pug-loader": "^2.4.0",
+    "ts-node": "~8.0.1",
+    "tslint": "~5.12.1",
+    "typescript": "~3.2.4"
   }
 }
