@@ -1,6 +1,6 @@
 import {Injectable} from "@angular/core";
 import {Http} from '@angular/http';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 
 import {BaseService} from "./base.service";
 
@@ -22,28 +22,28 @@ export class GroupService extends BaseService {
     }
 
     index() {
-        let groups = this.http.get(this.url(), this.options()).map(res => res.json());
+        let groups = this.http.get(this.url(), this.options()).pipe(map(res => res.json()));
         return groups;
     }
 
     get(id: string) {
-        let platform = this.http.get(this.url() + '/' + id, this.options()).map(res => res.json());
+        let platform = this.http.get(this.url() + '/' + id, this.options()).pipe(map(res => res.json()));
         return platform;
     }
 
 
     create(group: Group) {
-        let obs = this.http.post(this.url(), { 'group': group }, this.options()).map(res => res.json());
+        let obs = this.http.post(this.url(), { 'group': group }, this.options()).pipe(map(res => res.json()));
         return obs;
     }
 
 	update(group: Group) {
-		let obs = this.http.put(this.url() + '/' + group.id, { 'group': group }, this.options()).map(res => res.json());
+		let obs = this.http.put(this.url() + '/' + group.id, { 'group': group }, this.options()).pipe(map(res => res.json()));
         return obs;
 	}
 
 	delete(group: Group) {
-		let obs = this.http.delete(this.url() + '/' + group.id, this.options()).map(res => res.json());
+		let obs = this.http.delete(this.url() + '/' + group.id, this.options()).pipe(map(res => res.json()));
         return obs;
 	}
 }
