@@ -1,6 +1,6 @@
 import {Injectable} from "@angular/core";
 import {Http} from '@angular/http';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 
 import {BaseService} from "./base.service";
 
@@ -22,28 +22,28 @@ export class LicenseService extends BaseService {
     }
 
     index() {
-        let licenses = this.http.get(this.url(), this.options()).map(res => res.json());
+        let licenses = this.http.get(this.url(), this.options()).pipe(map(res => res.json()));
         return licenses;
     }
 
     get(id: string) {
-        let platform = this.http.get(this.url() + '/' + id, this.options()).map(res => res.json());
+        let platform = this.http.get(this.url() + '/' + id, this.options()).pipe(map(res => res.json()));
         return platform;
     }
 
 
     create(license: License) {
-        let obs = this.http.post(this.url(), { 'license': license }, this.options()).map(res => res.json());
+        let obs = this.http.post(this.url(), { 'license': license }, this.options()).pipe(map(res => res.json()));
         return obs;
     }
 
     update(license: License) {
-        let obs = this.http.put(this.url() + '/' + license.id, { 'license': license }, this.options()).map(res => res.json());
+        let obs = this.http.put(this.url() + '/' + license.id, { 'license': license }, this.options()).pipe(map(res => res.json()));
         return obs;
     }
 
     delete(license: License) {
-        let obs = this.http.delete(this.url() + '/' + license.id, this.options()).map(res => res.json());
+        let obs = this.http.delete(this.url() + '/' + license.id, this.options()).pipe(map(res => res.json()));
         return obs;
     }
 }
