@@ -1,6 +1,6 @@
 import {Component, Output, Inject} from '@angular/core';
 import {User} from '../models/user';
-import {Group} from '../models/Group';
+import {Group} from '../models/group';
 import {Status} from '../models/status';
 import {Search} from '../models/search';
 import {IdentityProvider} from '../models/identity_provider';
@@ -9,7 +9,7 @@ import {UUID} from 'angular2-uuid';
 
 import {ToasterModule, ToasterService} from 'angular2-toaster/angular2-toaster';
 
-import {SlideComponent, CarouselComponent, CarouselModule} from 'ng2-bootstrap';
+// import {SlideComponent, CarouselComponent, CarouselModule} from 'ng2-bootstrap';
 
 import {UserService} from '../services/user.service';
 import {GroupService} from '../services/group.service';
@@ -17,7 +17,7 @@ import {MarketplaceService} from '../services/marketplace.service';
 
 @Component({
     selector: 'groups',
-    templateUrl: '/groups.html'
+    templateUrl: '../views/groups.pug'
 })
 export class GroupsComponent {
 
